fix(cache): #4438 디스크 캐시 카운터를 usize atomic 으로 — 32-bit 타겟 빌드 회귀
graph.disk_load_hits / disk_cache.put_seq 가 Atomic(u64) 라 32-bit 타겟(wasm32-wasi,
x86-windows-msvc)에서 @atomicRmw 가 64-bit 정수를 거부 → atomic.zig:53 컴파일 실패
(CI 의 WASM Build / Publish Smoke / NAPI win32-ia32 job 동시 실패).

워드 크기 usize atomic 으로 전환: 32-bit=u32(atomic 지원), 64-bit=u64(기존과 동일),
단조 카운터/hit 카운터 의미 유지. minify.zig dead_toplevel_count 와 동일 선례.

+++ src/bundler/graph.zig
@@ -149,7 +149,8 @@ pub const ModuleGraph = struct {
     disk_load_enabled: bool = false,
     /// #4438 디스크 캐시 load hit 카운터(계측/테스트). parseModule 이 worker 에서 병렬 increment 하므로
     /// atomic. hit = parse 를 스킵하고 캐시에서 복원한 모듈 수.
-    disk_load_hits: std.atomic.Value(u64) = .init(0),
+    /// usize: 32-bit 타겟(wasm32/x86-win)은 64-bit atomic 을 지원하지 않으므로 워드 크기 카운터 사용.
+    disk_load_hits: std.atomic.Value(usize) = .init(0),
     /// perf/hmr-graph-topology-reuse Phase B — 위상(topology) 보존 모드.
     /// true 면 (1) `transferModulesToStore` 가 비활성(graph 가 parse_arena 단독 owner —
     /// store 로 양도하지 않음), (2) bundler external_graph 훅이 매 빌드 graph 를 전량 clear
